+++ src/views/PriceValidationAssistant.vue
@@ -49,6 +49,8 @@ export default {
   },
   data() {
     return {
+      currentDateTime: utils.currentDateTime(),  // usefull to avoid fetching duplicates during pagination
+      // data
       priceTagList: [],
       priceTagTotal: null,
       priceTagPage: 0,  // issue with pagination once the user starts removing/validating price tags...
@@ -72,7 +74,7 @@ export default {
       return 6
     },
     getPriceTagsParams() {
-      return { proof__owner: this.username, proof__ready_for_price_tag_validation: true, status__isnull: true, order_by: this.currentOrder, size: this.getApiSize, page: this.priceTagPage }
+      return { proof__owner: this.username, proof__ready_for_price_tag_validation: true, status__isnull: true, created__lte: this.currentDateTime, order_by: this.currentOrder, size: this.getApiSize, page: this.priceTagPage }
     },
   },
   mounted() {
@@ -127,29 +129,32 @@ export default {
           this.priceTagTotal = data.total
           this.loading = false
           for (let i = 0; i < data.items.length; i++) {
-            const label = data.items[i]['predictions'][0]['data']
-            const barcodeString = label.barcode ? utils.cleanBarcode(label.barcode.toString()) : ''
-            // TODO: some of these will be None if gemini did not give a proper reply, so detection and error handling is needed
-            const productPriceForm = {
-              id: data.items[i].id,
-              type: barcodeString.length >= 8 ? constants.PRICE_TYPE_PRODUCT : constants.PRICE_TYPE_CATEGORY,
-              category_tag: label.product,
-              origins_tags: [label.origin],
-              labels_tags: label.organic ? [constants.PRODUCT_CATEGORY_LABEL_ORGANIC] : [],
-              price: label.price.toString(),
-              price_per: label.unit,
-              price_is_discounted: false,
-              currency: data.items[i]['proof'].currency || this.appStore.getUserLastCurrencyUsed,
-              proof: data.items[i]['proof'],
-              proofImage: data.items[i]['proof'].file_path,
-              // proofImage: 'https://prices.openfoodfacts.org/img/0024/2NToLMxOgN.webp',
-              croppedImage: null,
-              product_code: barcodeString,
-              detected_product_code: barcodeString,
-              product_name: label.product_name,
-              bounding_box: data.items[i].bounding_box
+            // only validate price tags with predictions
+            if (data.items[i]['predictions'].length > 0) {
+              const label = data.items[i]['predictions'][0]['data']
+              const barcodeString = label.barcode ? utils.cleanBarcode(label.barcode.toString()) : ''
+              // TODO: some of these will be None if gemini did not give a proper reply, so detection and error handling is needed
+              const productPriceForm = {
+                id: data.items[i].id,
+                type: barcodeString.length >= 8 ? constants.PRICE_TYPE_PRODUCT : constants.PRICE_TYPE_CATEGORY,
+                category_tag: label.product,
+                origins_tags: [label.origin],
+                labels_tags: label.organic ? [constants.PRODUCT_CATEGORY_LABEL_ORGANIC] : [],
+                price: label.price.toString(),
+                price_per: label.unit,
+                price_is_discounted: false,
+                currency: data.items[i]['proof'].currency || this.appStore.getUserLastCurrencyUsed,
+                proof: data.items[i]['proof'],
+                proofImage: data.items[i]['proof'].file_path,
+                // proofImage: 'https://prices.openfoodfacts.org/img/0024/2NToLMxOgN.webp',
+                croppedImage: null,
+                product_code: barcodeString,
+                detected_product_code: barcodeString,
+                product_name: label.product_name,
+                bounding_box: data.items[i].bounding_box
+              }
+              this.productPriceForms.push(productPriceForm)
             }
-            this.productPriceForms.push(productPriceForm)
           }
         })
     },